Move around some code regarding message catalogs
