+++ cjworkbench/i18n/catalogs/merge.py
@@ -3,7 +3,6 @@ from cjworkbench.i18n import default_locale, supported_locales
 from cjworkbench.i18n.catalogs import (
     catalog_path,
     CATALOG_FILENAME,
-    BACKUP_CATALOG_FILENAME,
     TEMPLATE_CATALOG_FILENAME,
     COMMENT_TAG_FOR_DEFAULT_MESSAGE,
 )
@@ -11,6 +10,10 @@ import re
 import sys
 from os import remove
 from shutil import copyfile
+
+
+TEMP_BACKUP_CATALOG_FILENAME = "old.po"
+
 
 _default_message_re = re.compile(
     r"\s*" + COMMENT_TAG_FOR_DEFAULT_MESSAGE + ":\s*(.*)\s*"
@@ -39,7 +42,7 @@ def prepare():
         if locale != default_locale:
             copyfile(
                 catalog_path(locale, CATALOG_FILENAME),
-                catalog_path(locale, BACKUP_CATALOG_FILENAME),
+                catalog_path(locale, TEMP_BACKUP_CATALOG_FILENAME),
             )
 
 
@@ -59,7 +62,7 @@ def _merge_catalog(locale, source_catalog, default={}):
         catalog = read_po(target_catalog_file)
     if not default:
         with open(
-            catalog_path(locale, BACKUP_CATALOG_FILENAME), "r"
+            catalog_path(locale, TEMP_BACKUP_CATALOG_FILENAME), "r"
         ) as target_catalog_file:
             old = read_po(target_catalog_file)
     for message in source_catalog:
@@ -110,7 +113,7 @@ def clean():
     # Remove temp files for non-default locales
     for locale in supported_locales:
         if locale != default_locale:
-            remove(catalog_path(locale, BACKUP_CATALOG_FILENAME))
+            remove(catalog_path(locale, TEMP_BACKUP_CATALOG_FILENAME))
 
     # Update template file for default locale
     with open(catalog_path(default_locale, CATALOG_FILENAME)) as catalog_file:
